✨ onPress action on every ComposableScreen UIElement — SDK v1.37.0
Add optional `onPress: ButtonAction[]` to BaseBoxProps so any non-pressable
element can dispatch the same action list as Button (continue/setVariable/custom),
run sequentially. Wired via a single central Pressable in renderElement (mirroring
the AnimatedBox wrapper); the action-dispatch loop is extracted from ButtonElement
into a shared runActions helper. ButtonAction types moved to common.types.ts
(headless) / actions.ts (UI), re-exported from ButtonElement for back-compat.

Skipped for elements that own their own tap/focus/scroll gesture:
Button, RadioGroup, CheckboxGroup, DatePicker, Input, WheelPicker.
