+++ claude-plugin/skills/validate-step-json/SKILL.md
@@ -45,7 +45,8 @@ Run: `npx tsx scripts/_validate-composable.ts "$(cat step.json)"`
    - `RadioGroup.props.items` / `CheckboxGroup.props.items` is `[{label, value}]` (NOT `options`)
    - `WheelPicker.props` provides exactly one of `items: [{label, value}]` (unique values) or `range: {min, max, step?, unit?}` — both or neither is a schema error; `defaultValue` (if present) must match an available item value
    - `ProgressIndicator.props.variant` (if present) is `"linear"|"circular"`; `easing` (if present) is `"linear"|"ease-in"|"ease-out"|"ease-in-out"`; `value`/`initialValue` are 0–100; `duration`/`delay`/`thickness`/`size` are non-negative numbers; `autoplay`/`loop`/`showLabel` are booleans. Not a container — must NOT have `children`
-   - `Button.props.actions` is an array; entries are `"continue"` or `{type:"custom",function,variables?}` or `{type:"setVariable",name,value,valueMode?,kind?}` (`valueMode` ∈ `"literal"|"expression"`; `kind` ∈ `"int"|"float"|"string"` — all three action shapes are in the headless schema, `ButtonElement.ts`)
+   - `Button.props.actions` is an array; entries are `"continue"` or `{type:"custom",function,variables?}` or `{type:"setVariable",name,value,valueMode?,kind?}` (`valueMode` ∈ `"literal"|"expression"`; `kind` ∈ `"int"|"float"|"string"` — all three action shapes are the `ButtonAction` union in the headless schema, `common.types.ts`, re-exported from `ButtonElement.ts`)
+   - `props.onPress` (any element — it's a `BaseBoxProps` field) is a `ButtonAction[]` with the same entry shapes as `Button.props.actions`. Schema-valid on every element; runtime ignores it on the elements that own their own gesture (`Button` — use `actions`; `RadioGroup`, `CheckboxGroup`, `DatePicker`, `Input`, `WheelPicker`) — warn if seen there, error only on a malformed action entry
    - `Button.props.disabledWhen` (NOT `disabled`) is a valid `LeafCondition` or `ConditionGroup`
    - **Gating sanity (warning, not schema error)**: a `disabledWhen` / `renderWhen` condition should reference a variable captured on the SAME screen (this screen's `variableName` elements or `setVariable` actions) or a deliberate upstream capture. A CTA gated on a variable this screen never touches is almost always a copy-paste bug (seen in production: picker CTAs gated on a stale `goals` variable from an earlier screen) — flag it.
    - `Button.props.pressedStyle` / `disabledStyle` (if present) are objects — `Partial` of overridable Button props; must NOT nest `pressedStyle`/`disabledStyle`. `transitionDurationMs` is a non-negative number.
